refactor: extract permission request manager from session
